Added "names" setting to context
This replaces "characters" as it is more generic.

Moved description and names into the prompt message context - it doesn't seem to make a difference having them in a separate message.

Updated instructions to encourage use of provided names in translation. gpt-3.5-turbo isn't really smart enough to handle it, but interestingly gpt3.5-turbo-instruct more or less is.